CM Metadata
Since both DS and Configuration Admin parse the cm section in the manifest.json file this code was moved from the  Configuration Admin namespace to the cppmicroservices/util namespace. Signed-off-by : The MathWorks, Inc. <pelliott@mathworks.com>
