+++ compendium/ConfigurationAdmin/CMakeLists.txt
@@ -53,6 +53,8 @@ target_include_directories(ConfigurationAdmin PRIVATE
   ${CppMicroServices_BINARY_DIR}/compendium/LogService/include
   ${CppMicroServices_SOURCE_DIR}/compendium/CM/include
   ${CppMicroServices_BINARY_DIR}/compendium/CM/include
+  ${CppMicroServices_SOURCE_DIR}/cppmicroservices/util/include
+  ${CppMicroServices_BINARY_DIR}/cppmicroservices/util/include
   ${CppMicroServices_SOURCE_DIR}/third_party/googletest/googletest/include
   ${CppMicroServices_SOURCE_DIR}/third_party/googletest/googlemock/include
   )
