+++ compendium/ConfigurationAdmin/test/TestMetadataParserFactory.cpp
@@ -21,8 +21,8 @@
  =============================================================================*/
 #include "gtest/gtest.h"
 
-#include "../src/metadata/MetadataParserFactory.hpp"
-#include "../src/metadata/MetadataParser.hpp"
+#include "cppmicroservices/util/MetadataParserFactory.hpp"
+#include "cppmicroservices/util/MetadataParser.hpp"
 #include "Mocks.hpp"
 
 namespace cppmicroservices {
@@ -32,11 +32,11 @@ namespace cppmicroservices {
       TEST(TestMetadataParserFactory, ManifestVersionInvalid) {
         auto logger = std::make_shared<FakeLogger>();
         EXPECT_THROW(
-        MetadataParserFactory::Create(0, logger);,
+        cppmicroservices::util::MetadataParserFactory::Create(0, logger);,
         std::runtime_error);
         EXPECT_THROW(
-        MetadataParserFactory::Create(2, logger);,
-        std::runtime_error);
+        cppmicroservices::util::MetadataParserFactory::Create(2, logger);
+          , std::runtime_error);
       }
     }
   }
