+++ compendium/DeclarativeServices/CMakeLists.txt
@@ -55,10 +55,16 @@ include_directories(${CppMicroServices_SOURCE_DIR}/framework/include
   ${CppMicroServices_BINARY_DIR}/framework/include
   ${CppMicroServices_BINARY_DIR}/compendium/ServiceComponent/include
   ${CppMicroServices_BINARY_DIR}/compendium/LogService/include
+  ${CppMicroServices_BINARY_DIR}/compendium/CM/include
+  ${CppMicroServices_BINARY_DIR}/cppmicroservices/cm/include
+  ${CppMicroServices_BINARY_DIR}/util/include
+  ${CppMicroServices_SOURCE_DIR}/util/include
   ${CppMicroServices_SOURCE_DIR}/compendium/ServiceComponent/include
   ${CppMicroServices_SOURCE_DIR}/compendium/LogService/include
+  ${CppMicroServices_SOURCE_DIR}/compendium/CM/include
   ${CppMicroServices_SOURCE_DIR}/third_party/googletest/googletest/include
   ${CppMicroServices_SOURCE_DIR}/third_party/googletest/googlemock/include
+  ${CppMicroServices_SOURCE_DIR}/cppmicroservices/cm/include
   )
 
 # There are warnings in the boost asio headers which are flagged as errors. Include the boost
