+++ util/CMakeLists.txt
@@ -10,12 +10,19 @@ add_library(util OBJECT
   include/cppmicroservices/util/MappedFile.h
   include/cppmicroservices/util/ScopeGuard.h
   include/cppmicroservices/util/String.h
+  include/cppmicroservices/util/CMConstants.hpp
+  include/cppmicroservices/util/ConfigurationMetadata.hpp
+  include/cppmicroservices/util/MetadataParser.hpp
+  include/cppmicroservices/util/MetadataParserFactory.hpp
+  include/cppmicroservices/util/MetadataParserImpl.hpp
 
   src/BundleObjFactory.cpp
   src/BundleObjFile.cpp
   src/Error.cpp
   src/FileSystem.cpp
   src/String.cpp
+  src/CMConstants.cpp
+  src/MetadataParserImpl.cpp
 )
 
 set_property(TARGET util PROPERTY POSITION_INDEPENDENT_CODE 1)
@@ -23,7 +30,14 @@ set_property(TARGET util PROPERTY POSITION_INDEPENDENT_CODE 1)
 target_include_directories(util
   PUBLIC
     ${CMAKE_CURRENT_SOURCE_DIR}/include
+    ${CppMicroServices_BINARY_DIR}/framework/include
+    ${CppMicroServices_SOURCE_DIR}/framework/include
+    ${CppMicroServices_SOURCE_DIR}/compendium/LogService/include
+    ${CppMicroServices_BINARY_DIR}/compendium/LogService/include
+ 
   PRIVATE
     ${PROJECT_SOURCE_DIR}/third_party
     ${PROJECT_BINARY_DIR}/include
 )
+
+add_dependencies(util usLogService)
